Review + fix errors
