--- src/CATEGORIE.java
@@ -1,8 +1,0 @@
-package src;
-
-public enum CATEGORIE {
-        LACTATE,
-        PATISERIE,
-        MEZELURI,
-        ALTELE
-    }
