+++ src/produse/tipuri/Categorie.java
@@ -1,0 +1,8 @@
+package src.produse.tipuri;
+
+public enum Categorie {
+    LACTATE,
+    PATISERIE,
+    MEZELURI,
+    ALTELE
+}
